Refactor embedded core and adapter utilities
