+++ packages/adapter-utils/tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "extends": "@bullstudio/typescript-config/base.json",
+  "include": ["src"],
+  "exclude": ["node_modules", "src/**/*.test.ts"],
+  "compilerOptions": {
+    "outDir": "dist",
+    "rootDir": "src"
+  }
+}
