+++ packages/bull-adapter/package.json
@@ -17,6 +17,7 @@
     "bull": "^4.0.0"
   },
   "dependencies": {
+    "@bullstudio/adapter-utils": "workspace:*",
     "@bullstudio/connect-types": "workspace:*",
     "@bullstudio/embedded-core": "workspace:*"
   },
