+++ packages/bull-adapter/src/index.ts
@@ -1,3 +1,11 @@
+import {
+  createJobNotFoundError,
+  createWorkerCount,
+  filterJobsByName,
+  normalizeJobCounts,
+  sortJobs,
+  toJobSummary,
+} from "@bullstudio/adapter-utils";
 import type { Job, JobCounts, JobStatus } from "@bullstudio/connect-types";
 import type { QueueAdapter } from "@bullstudio/embedded-core";
 import type Bull from "bull";
@@ -57,9 +65,7 @@ export function createBullQueueAdapter(
         .filter((job): job is Bull.Job => job !== undefined && job !== null)
         .map((job) => mapJob(job, queue.name));
 
-      if (filter?.name) {
-        mappedJobs = mappedJobs.filter((job) => job.name === filter.name);
-      }
+      mappedJobs = filterJobsByName(mappedJobs, filter?.name);
 
       if (sort) {
         mappedJobs = sortJobs(mappedJobs, sort.field, sort.order);
@@ -74,10 +80,17 @@ export function createBullQueueAdapter(
         (options?.offset ?? 0) + (options?.limit ?? 100) - 1,
       );
 
-      return jobs
+      let summaries = jobs
         .filter((job): job is Bull.Job => job !== undefined && job !== null)
-        .map((job) => mapJob(job, queue.name))
-        .map(({ data, returnValue, stacktrace, ...summary }) => summary);
+        .map((job) => mapJob(job, queue.name));
+
+      summaries = filterJobsByName(summaries, options?.filter?.name);
+
+      if (options?.sort) {
+        summaries = sortJobs(summaries, options.sort.field, options.sort.order);
+      }
+
+      return summaries.map(toJobSummary);
     },
     getJob: async (jobId) => {
       const job = await queue.getJob(jobId);
@@ -87,27 +100,20 @@ export function createBullQueueAdapter(
     retryJob: async (jobId) => {
       const job = await queue.getJob(jobId);
       if (!job) {
-        throw new Error(
-          `Job "${jobId}" was not found in queue "${queue.name}".`,
-        );
+        throw createJobNotFoundError(queue.name, jobId);
       }
       await job.retry();
     },
     removeJob: async (jobId) => {
       const job = await queue.getJob(jobId);
       if (!job) {
-        throw new Error(
-          `Job "${jobId}" was not found in queue "${queue.name}".`,
-        );
+        throw createJobNotFoundError(queue.name, jobId);
       }
       await job.remove();
     },
     getWorkerCount: async () => {
       const workers = await queue.getWorkers();
-      return {
-        queueName: queue.name,
-        count: workers.length,
-      };
+      return createWorkerCount(queue.name, workers);
     },
   };
 }
@@ -174,42 +180,10 @@ function normalizeProgress(progress: number | object): number | object {
   return 0;
 }
 
-function sortJobs(
-  jobs: Job[],
-  field: "timestamp" | "processedOn" | "finishedOn" | "progress",
-  order: "asc" | "desc",
-): Job[] {
-  return [...jobs].sort((a, b) => {
-    const aValue =
-      field === "progress"
-        ? typeof a.progress === "number"
-          ? a.progress
-          : 0
-        : (a[field] ?? 0);
-    const bValue =
-      field === "progress"
-        ? typeof b.progress === "number"
-          ? b.progress
-          : 0
-        : (b[field] ?? 0);
-
-    return order === "asc" ? aValue - bValue : bValue - aValue;
-  });
-}
-
 async function getJobCounts(queue: Bull.Queue): Promise<JobCounts> {
   const counts = await queue.getJobCounts();
 
-  return {
-    waiting: counts.waiting ?? 0,
-    active: counts.active ?? 0,
-    completed: counts.completed ?? 0,
-    failed: counts.failed ?? 0,
-    delayed: counts.delayed ?? 0,
-    paused: 0,
-    prioritized: 0,
-    waitingChildren: 0,
-  };
+  return normalizeJobCounts(counts);
 }
 
 function getQueuePrefix(queue: Bull.Queue): string {
