+++ packages/bullmq-adapter/src/index.test.ts
@@ -124,6 +124,59 @@ describe("createBullMqQueueAdapter", () => {
     ]);
   });
 
+  it("applies name filtering and sorting when reading job summaries", async () => {
+    const queue = {
+      name: "email",
+      getJobs: async () => [
+        createQueueJob({ id: "1", name: "digest", timestamp: 300 }),
+        createQueueJob({ id: "2", name: "welcome", timestamp: 100 }),
+        createQueueJob({ id: "3", name: "welcome", timestamp: 200 }),
+      ],
+    } as unknown as Queue;
+
+    const adapter = createBullMqQueueAdapter(queue);
+
+    await expect(
+      adapter.getJobsSummary({
+        filter: { name: "welcome" },
+        sort: { field: "timestamp", order: "desc" },
+      }),
+    ).resolves.toEqual([
+      {
+        id: "3",
+        name: "welcome",
+        queueName: "email",
+        status: "waiting",
+        progress: 0,
+        attemptsMade: 0,
+        failedReason: undefined,
+        timestamp: 200,
+        processedOn: undefined,
+        finishedOn: undefined,
+        delay: undefined,
+        priority: undefined,
+        parentId: undefined,
+        repeatJobKey: undefined,
+      },
+      {
+        id: "2",
+        name: "welcome",
+        queueName: "email",
+        status: "waiting",
+        progress: 0,
+        attemptsMade: 0,
+        failedReason: undefined,
+        timestamp: 100,
+        processedOn: undefined,
+        finishedOn: undefined,
+        delay: undefined,
+        priority: undefined,
+        parentId: undefined,
+        repeatJobKey: undefined,
+      },
+    ]);
+  });
+
   it("delegates job and queue operations without closing the supplied queue", async () => {
     const retry = vi.fn<() => Promise<void>>();
     const remove = vi.fn<() => Promise<void>>();
@@ -180,3 +233,23 @@ describe("createBullMqQueueAdapter", () => {
     expect(adapter).not.toHaveProperty("disconnect");
   });
 });
+
+function createQueueJob(overrides: {
+  id: string;
+  name: string;
+  timestamp: number;
+}) {
+  return {
+    data: { userId: 123 },
+    progress: 0,
+    attemptsMade: 0,
+    opts: {},
+    failedReason: undefined,
+    stacktrace: [],
+    returnvalue: { ok: true },
+    processedOn: undefined,
+    finishedOn: undefined,
+    getState: async () => "waiting",
+    ...overrides,
+  };
+}
