+++ packages/embedded-core/src/errors.ts
@@ -1,0 +1,6 @@
+export class ReadOnlyDashboardError extends Error {
+  constructor() {
+    super("Read-only dashboards cannot mutate queues or jobs.");
+    this.name = "ReadOnlyDashboardError";
+  }
+}
