+++ packages/embedded-core/src/mutation.ts
@@ -1,0 +1,16 @@
+import { ReadOnlyDashboardError } from "./errors";
+import type { ResolvedDashboardConfig } from "./types";
+
+export async function withMutationAccess<T>(
+  config: ResolvedDashboardConfig,
+  operation: () => Promise<T>,
+): Promise<T> {
+  assertCanMutate(config);
+  return operation();
+}
+
+export function assertCanMutate(config: ResolvedDashboardConfig): void {
+  if (config.readOnly) {
+    throw new ReadOnlyDashboardError();
+  }
+}
